refactor: extract ReactivityEngine from builder mixin chain
Reactivity (subscribe, formula re-execution, timers, incremental
compile, output suspension) is now encapsulated in ReactivityEngine,
a standalone class created lazily by build(). The builder base class
has only 3 mixin: _GrammarMixin, _BuildMixin, _OutputMixin.

ReactivityEngine holds its own state (binding, formula registry,
timers) and calls back into the builder for build operations.
Without subscribe(), the engine exists but is not activated.

Remove _reactivity_mixin.py (absorbed into _reactivity.py).

583 tests pass, ruff clean, coverage 90%.
